fixup! fixup! add tx/rx metrics for peers
move peermetrics into separate TUs
